+++ resources/lib/app/__init__.py
@@ -9,12 +9,14 @@ from .application import App
 from .connection import Connection
 from .libsync import Sync
 from .playstate import PlayState
+from .playqueues import Playqueues
 
 ACCOUNT = None
 APP = None
 CONN = None
 SYNC = None
 PLAYSTATE = None
+PLAYQUEUES = None
 
 
 def init(entrypoint=False):
@@ -22,13 +24,15 @@ def init(entrypoint=False):
     entrypoint=True initiates only the bare minimum - for other PKC python
     instances
     """
-    global ACCOUNT, APP, CONN, SYNC, PLAYSTATE
+    global ACCOUNT, APP, CONN, SYNC, PLAYSTATE, PLAYQUEUES
     APP = App(entrypoint)
     CONN = Connection(entrypoint)
     ACCOUNT = Account(entrypoint)
     SYNC = Sync(entrypoint)
     if not entrypoint:
         PLAYSTATE = PlayState()
+        PLAYQUEUES = Playqueues()
+
 
 def reload():
     """
